Refactor: "Exception 열거형 패키지 변경"
