+++ src/main/java/org/library/exception/AlreadyLostException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class AlreadyLostException extends RuntimeException {
 
     public AlreadyLostException() {
