+++ src/main/java/org/library/exception/AlreadyOrganizingException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class AlreadyOrganizingException extends RuntimeException {
 
     public AlreadyOrganizingException() {
