+++ src/main/java/org/library/exception/AlreadyRentException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class AlreadyRentException extends RuntimeException {
 
     public AlreadyRentException() {
