+++ src/main/java/org/library/exception/InvalidFuncException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class InvalidFuncException extends RuntimeException {
 
     public InvalidFuncException() {
