+++ src/main/java/org/library/exception/NotAvailableException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class NotAvailableException extends RuntimeException {
 
     public NotAvailableException() {
