+++ src/main/java/org/library/exception/NotExistException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class NotExistException extends RuntimeException {
 
     public NotExistException() {
