+++ src/main/java/org/library/exception/NotReturnsException.java
@@ -1,7 +1,5 @@
 package org.library.exception;
 
-import org.library.entity.Exception;
-
 public class NotReturnsException extends RuntimeException {
 
     public NotReturnsException() {
